--- a/nemo/lightning/pytorch/callbacks/__init__.py
+++ b/nemo/lightning/pytorch/callbacks/__init__.py
@@ -19,6 +19,7 @@ from nemo.lightning.pytorch.callbacks.memory_profiler import MemoryProfileCallba
 from nemo.lightning.pytorch.callbacks.model_checkpoint import ModelCheckpoint
 from nemo.lightning.pytorch.callbacks.model_transform import ModelTransform
 from nemo.lightning.pytorch.callbacks.nsys import NsysCallback
+from nemo.lightning.pytorch.callbacks.workload_inspector import WorkloadInspectorCallback
 from nemo.lightning.pytorch.callbacks.peft import PEFT
 from nemo.lightning.pytorch.callbacks.preemption import PreemptionCallback
 from nemo.lightning.pytorch.callbacks.progress_bar import MegatronProgressBar
@@ -30,6 +31,7 @@ __all__ = [
     "ModelTransform",
     "PEFT",
     "NsysCallback",
+    "WorkloadInspectorCallback",
     "MegatronProgressBar",
     "ProgressPrinter",
     "PreemptionCallback",